Fix Lazy management window background highlights
The Lazy management window looked out of place next to Toggleterm and
Telescope windows. This commit gives it a different background color.

Note that the background color doesn't "cover" the entire Lazy window,
thats why the "Normal" highlight was chosen. It now has the same bg
color as regular buffers.